+++ CHANGELOG.md
@@ -13,3 +13,10 @@ Initial release.  Provides basic data source functionality of fetching saved/sch
 
 - Remove unnecessary logic from CheckHealth
 - Properly close HTTP response body in all cases
+
+## 1.1.2
+
+- Query timeout is now configurable.
+- Add info & doc link to the timeout error message to encourage the user to leverage scheduled search.
+- Upon timeout, cancel the query on the Cribl Search side.
+- Allow saved search ID to be composed using dashboard `${variables}`.